Add script(s) to automate template creation
The scripts are almost fully automatic. There is one additional step the
operator needs to do beyond just launch the script.

The output of the new main scipt is a *.vma.zst file that can be
updated to the cm server and downloaded by students when they run an
updated install-templates.sh script.

Signed-off-by: John Bonesio <bones@bonesio.net>